Create F1_UDP_Client.py
Listens to the desired ip and port for the F1 udp packet stream. In this case, the mini server we created to test out coming data. In theory, when finished to accept all the packets F1 2018 sends out, should work for that just by specifying the correct ip and port!