Added pcap emulator
Emulates sending a packets captured from codemasters f1 2018 racing game that where captured into a .PCAP file

Currently does not emulate the time at which they were sent. Simply just sends out in brute force.

Will try to emulate the time and frequency for future improvements later.

Is used to test this work when not in front of and running codemasters f1 2018 game